// src/modules/user/constants/residencePeriod.enum.ts
export enum ResidencePeriod {
  SIX_MONTHS_UNDER = '6_MONTHS_UNDER', // 6개월 미만
  SIX_MONTHS_TO_3_YEARS = '6_MONTHS_TO_3_YEARS', // 6개월 이상 ~ 3년 미만
  THREE_YEARS_OVER = '3_YEARS_OVER', // 3년 이상~
  PLANNED_MOVE = 'PLANNED_MOVE', // 이주 예정
  SHORT_TERM_STAY = 'SHORT_TERM_STAY' // 단기 체류 중
}